% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucls.R
\name{add_fe_complpool}
\alias{add_fe_complpool}
\title{Add fe (factor_exceeding) based on complete-pooling UCL to data frame}
\usage{
add_fe_complpool(d)
}
\arguments{
\item{d}{Initialized data frame with UCLs}
}
\value{
Data frame with values for \emph{fe_complpool}
}
\description{
Adds fe (factor_exceeding) based on complete-pooling UCL to data frame,
given \emph{n_new} and complete-pooling estimate of \emph{theta}.
}
\details{
\emph{fe_complpool}: factor \emph{f}, with observed \emph{y_new}
exceeding complete-pooling UCL by \emph{f * sd(y_new)},
given \emph{n_new} and complete-pooling estimate of \emph{theta}
}
