% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucls.R
\name{add_fe_nopool}
\alias{add_fe_nopool}
\title{Add fe (factor_exceeding) based on no-pooling UCL to data frame}
\usage{
add_fe_nopool(d)
}
\arguments{
\item{d}{Initialized data frame with UCLs}
}
\value{
Data frame with values for \emph{fe_nopool}
}
\description{
Adds fe (factor_exceeding) based on no-pooling UCL to data frame, given
\emph{n_new} and no-pooling estimate of \emph{theta}.
}
\details{
\emph{fe_nopool}: factor \emph{f}, with observed y_new
exceeding no-pooling UCL by \eqn{f * sd(y_new)},
given n_new and no-pooling estimate of theta
}
