% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucls.R
\name{add_fe_partpool}
\alias{add_fe_partpool}
\title{Add fe (factor_exceeding) based on partial-pooling UCL to data frame}
\usage{
add_fe_partpool(d)
}
\arguments{
\item{d}{Initialized data frame with UCLs}
}
\value{
Data frame with values for \emph{fe_partpool}
}
\description{
Adds fe (factor_exceeding) based on partial-pooling UCL to data frame,
given \emph{n_new} and partial-pooling estimate of \emph{theta}.
}
\details{
\emph{fe_partpool}: factor \emph{f}, with observed y_new
exceeding partial-pooling UCL by \eqn{f * sd(y_new)},
given n_new and partial-pooling estimate of theta
}
