% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucls.R
\name{add_fe_true_theta}
\alias{add_fe_true_theta}
\title{Add fe (factor_exceeding) based on "true" UCL to data frame}
\usage{
add_fe_true_theta(d)
}
\arguments{
\item{d}{Initialized data frame with UCLs}
}
\value{
Data frame with values for \emph{fe_true_theta}
}
\description{
Adds fe (factor_exceeding) based on true UCL to data frame, given
\emph{n_new} and true value of \emph{theta}.
}
\details{
\emph{fe_true_theta}: factor \emph{f}, with observed y_new
exceeding true-theta UCL by \emph{f * sd(y_new)},
given n_new and true value of theta (if known)
}
