% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{add_theta_nopool}
\alias{add_theta_nopool}
\title{Add no-pooling estimate of \emph{theta} to data frame}
\usage{
add_theta_nopool(d)
}
\arguments{
\item{d}{Initialized data frame}
}
\value{
Data frame with values for \emph{theta_nopool}
}
\description{
Adds no-pooling estimate of \emph{theta} to data frame.
}
\keyword{internal}
