% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucls.R
\name{add_ucl_complpool}
\alias{add_ucl_complpool}
\title{Add complete-pooling UCL to data frame}
\usage{
add_ucl_complpool(d)
}
\arguments{
\item{d}{Initialized data frame}
}
\value{
data frame with values for \emph{ucl_complpool}
}
\description{
Add UCL based on \emph{n_new} and complete-pooling estimate of \emph{theta}
to data frame.
}
