% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucls.R
\name{add_ucl_partpool}
\alias{add_ucl_partpool}
\title{Add partial-pooling UCL (based on Bayesian hierarchical model) to data frame}
\usage{
add_ucl_partpool(d)
}
\arguments{
\item{d}{Initialized data frame}
}
\value{
data frame with values for ucl_partpool
}
\description{
UCL based on n_new and partial-pooling estimate of theta
}
