% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucls.R
\name{add_ucl_true_theta}
\alias{add_ucl_true_theta}
\title{Add "true" UCL to data frame}
\usage{
add_ucl_true_theta(d)
}
\arguments{
\item{d}{Initialized data frame}
}
\value{
data frame with values for ucl_true_theta
}
\description{
UCL based on n_new and true value of theta
}
