% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{complpool}
\alias{complpool}
\title{Complete-pooling model}
\usage{
complpool(d)
}
\arguments{
\item{d}{Initialized data frame}
}
\value{
data frame with values for theta_complpool
}
\description{
For the complete-pooling model we consider theta_i to be the same
for all observational units. We assume that observed data can be
treated as coming from a homogeneous source. For all units i our
complete-pooling estimate of theta_i is the ratio of the sum of
the observed count values of interest y_i to the sum of the
reference count values n_i.
}
