% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{countcheck}
\alias{countcheck}
\title{Return data frame combining data and results}
\usage{
countcheck(
  unit = NULL,
  n = NULL,
  y = NULL,
  n_new = NULL,
  y_new = NULL,
  true_theta = NULL,
  beta = 0.376,
  factor_sd = 3,
  random_seed = 200731,
  precis_depth = 1
)
}
\arguments{
\item{unit}{ID for unit}

\item{n}{Previous reference count values (measure of exposure),
must at least be 1}

\item{y}{Previous count values of interest}

\item{n_new}{New reference count values (measure of exposure),
must at least be 1}

\item{y_new}{New count values of interest}

\item{true_theta}{True value of theta (if known; optional)}

\item{beta}{Parameter for half-normal distribution of alpha
(default: 0.376, so that the expected value of alpha is 0.3) - used
in partial-pooling model}

\item{factor_sd}{Factor multiplying standard deviation (default: 3) - used
in calculation of UCLs}

\item{random_seed}{Seed value (default: 200731) - used in simulation and
by Stan}

\item{precis_depth}{Depth parameter for rethinking::precis() (default: 1) -
used by function for partial-pooling model}
}
\value{
Data frame with all columns populated
}
\description{
Constructs data frame with these fields:
\itemize{
\item \emph{unit}:            ID for unit
\item \emph{n}:               previous reference count values (measure of
                                exposure), must at least be 1
\item \emph{y}:               previous count values of interest
\item \emph{n_new}:           new reference count values (measure of
                                exposure), must at least be 1
\item \emph{y_new}:           new count values of interest
\item \emph{true_theta}:      true value of theta (if known)
\item \emph{theta_nopool}:    no-pooling estimate of theta
\item \emph{theta_complpool}: complete-pooling estimate of theta
\item \emph{theta_partpool}:  partial-pooling estimate of theta
\item \emph{ucl_true_theta}:  UCL based on n_new and true value of theta
                                (if known)
\item \emph{ucl_nopool}:      UCL based on n_new and no-pooling estimate
                                of theta
\item \emph{ucl_complpool}:   UCL based on n_new and complete-pooling
                                estimate of theta
\item \emph{ucl_partpool}:    UCL based on n_new and partial-pooling
                                estimate of theta
\item \emph{fe_true_theta}:   factor \emph{f}, with observed y_new
                                exceeding true-theta UCL by
                                \emph{f * sd(y_new)},
                                given n_new and true value of theta
                                (if known)
\item \emph{fe_nopool}:       factor \emph{f}, with observed y_new
                                exceeding no-pooling UCL by
                                \emph{f * sd(y_new)},
                                given n_new and no-pooling estimate
                                of theta
\item \emph{fe_complpool}:    factor \emph{f}, with observed y_new
                                exceeding complete-pooling UCL by
                                \emph{f * sd(y_new)},
                                given n_new and complete-pooling estimate
                                of theta
\item \emph{fe_partpool}:     factor \emph{f}, with observed y_new
                                exceeding partial-pooling UCL by
                                \emph{f * sd(y_new)},
                                given n_new and partial-pooling estimate
                                of theta
}
If data is provided via parameters, vectors must be of equal length,
with length >= 3. If \emph{unit} is NULL, all data vectors passed
to this function are ignored, and the complete data set is generated
in a simulation run.
}
