% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{escape}
\alias{escape}
\title{Escape special characters so that string can be inserted
into HTML}
\usage{
escape(string)
}
\arguments{
\item{string}{Input string}
}
\value{
String with special characters escaped
}
\description{
Escapes special characters so that string can be inserted
into HTML, in particular <, >, &, " and '.
}
\keyword{internal}
