% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_01}
\alias{example_01}
\title{Simulated data without anomalies}
\format{
A data frame with 1000 rows and 5 variables:
\itemize{
\item \emph{unit}:            ID for unit
\item \emph{n}:               previous reference count values (measure of
                                exposure), must at least be 1
\item \emph{y}:               previous count values of interest
\item \emph{n_new}:           new reference count values (measure of
                                exposure), must at least be 1
\item \emph{y_new}:           new count values of interest
}
}
\usage{
example_01
}
\description{
The first five columns of data frame created with
countcheck(random_seed = 200807)
}
\keyword{datasets}
