% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucls.R
\name{factor_exceeding}
\alias{factor_exceeding}
\title{Calculate by how many standard deviations y_new exceeds upper
control limit (UCL)}
\usage{
factor_exceeding(theta_hat, n_new, y_new, ucl)
}
\arguments{
\item{theta_hat}{Estimate (unless known) of rate parameter \emph{theta}}

\item{n_new}{New value for exposure (reference counts)}

\item{y_new}{New count values of interest}

\item{ucl}{Upper control limit (UCL)}
}
\value{
Factor \emph{e}, with observed \emph{y_new} exceeding UCL by
  \eqn{e * sd(y_new)}
}
\description{
Calculates factor \emph{e}, with observed \emph{y_new} exceeding UCL by
  \eqn{e * sd(y_new)}.
}
\details{
Assuming a Poisson distribution with
\deqn{lambda_hat = theta_hat * n_new}
the estimated standard deviation of \emph{y_new}
is
\deqn{sqrt(lambda_hat)}
}
