% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{html_body}
\alias{html_body}
\title{Generate HTML body}
\usage{
html_body(countcheck_list, unit_df, column_headers, home_url = NULL)
}
\arguments{
\item{countcheck_list}{List of data frame with columns
\emph{y_new}, \emph{ucl_partpool}, \emph{unit}}

\item{unit_df}{Data frame with columns
\emph{unit}, \emph{unit_group_name}, \emph{unit_name}, \emph{unit_url}}
}
\value{
String with HTML code for body
}
\description{
Generates HTML body.
}
\keyword{internal}
