% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{html_head}
\alias{html_head}
\title{Generate HTML head}
\usage{
html_head(title, charset, table_width_px, style)
}
\arguments{
\item{title}{Title for HTML document}

\item{charset}{Character set of data and HTML document}

\item{table_width_px}{Table width in pixels (for all tables)}

\item{style}{Additional CSS rules (optional)}
}
\value{
String with HTML code for head
}
\description{
Generates HTML head.
}
\keyword{internal}
