% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{html_report}
\alias{html_report}
\title{Generate HTML report}
\usage{
html_report(
  countcheck_list,
  unit_df,
  title = "Report",
  column_headers = c(group = "Group", count = "Count", ucl = "UCL", unit = "Unit", name
    = "Name"),
  charset = "utf-8",
  lang = "en",
  home_url = NULL,
  style = NULL
)
}
\arguments{
\item{countcheck_list}{List of data frames with columns
\emph{y_new}, \emph{ucl_partpool}, \emph{unit}}

\item{unit_df}{Data frame with columns
\emph{unit}, \emph{unit_group_name}, \emph{unit_name}, \emph{unit_url}}
}
\value{
String with HTML code for report
}
\description{
Generates HTML report.
}
