% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{html_table}
\alias{html_table}
\title{Generate HTML table}
\usage{
html_table(
  countcheck_df,
  unit_df,
  caption,
  headers = c(group = "Group", count = "Count", ucl = "UCL", unit = "Unit", name =
    "Name")
)
}
\arguments{
\item{countcheck_df}{Data frame with columns
\emph{y_new}, \emph{ucl_partpool}, and \emph{unit} -
as returned by \emph{select_for_report}}

\item{unit_df}{Data frame with columns
\emph{unit}, \emph{unit_group_name}, \emph{unit_name},
and \emph{unit_url} (ignored if NA).
Must contain exactly one record for each unit in
\emph{countcheck_df} dataframes in \emph{countcheck_list}.}

\item{caption}{Caption for table}

\item{headers}{Vector with five column headers -
the elements of the vector must be named \emph{group}, \emph{count},
\emph{ucl}, \emph{unit}, and \emph{name}}
}
\value{
String with HTML code for table
}
\description{
Generates HTML table.
}
\keyword{internal}
