% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{initialize}
\alias{initialize}
\title{Initialize data frame}
\usage{
initialize(n, y, n_new, y_new, true_theta = NULL)
}
\arguments{
\item{n}{Previous reference count values (measure of exposure)}

\item{y}{Previous count values of interest}

\item{n_new}{New reference count values}

\item{y_new}{New count values of interest}

\item{true_theta}{True value of theta (if known; optional)}
}
\value{
Initialized data frame
}
\description{
Construct data frame with these fields:
- unit:            index for unit
- n:               previous reference count values (measure of exposure)
- y:               previous count values of interest
- n_new:           new reference count values
- y_new:           new count values of interest
- true_theta:      true value of theta (if known)
- theta_nopool:    no-pooling estimate of theta
- theta_complpool: complete-pooling estimate of theta
- theta_partpool:  partial-pooling estimate of theta
- ucl_true_theta:  UCL based on n_new and true value of theta (if known)
- ucl_nopool:      UCL based on n_new and no-pooling estimate of theta
- ucl_complpool:   UCL based on n_new and complete-pooling estimate of theta
- ucl_partpool:    UCL based on n_new and partial-pooling estimate of theta
}
\details{
unit is initialized based on the length of \code{n}.
n, y, n_new, y_new, and true_theta (if known) are initialized using the
corresponding parameters. The other fields are initialized with NA.
}
