% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{matching_unit_group}
\alias{matching_unit_group}
\title{Selects rows with matching unit group}
\usage{
matching_unit_group(s, unit_df, unit_group_name)
}
\arguments{
\item{s}{Data frame of same type as returned by \emph{select_for_report()}}

\item{unit_df}{Data frame with columns
\emph{unit}, \emph{unit_group_name}, \emph{unit_name},
and \emph{unit_url} (ignored if NA).
Must contain exactly one record for each unit in
\emph{s}.}

\item{unit_group_name}{Name of unit group for which data
should be returned. Can be part of name (substring).}
}
\value{
Vector of type logical
}
\description{
Selects rows with matching unit group - helper function used
in select_for_report(). Condition: \emph{unit_group_name}
matches part of group name in \emph{unit_df} for units in \emph{s}.
}
\keyword{internal}
