% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{nopool}
\alias{nopool}
\title{No-pooling model}
\usage{
nopool(d)
}
\arguments{
\item{d}{Initialized data frame}
}
\value{
data frame with values for theta_nopool
}
\description{
For the no-pooling model we estimate theta_i using only the observed
counts of observational unit i. For each unit i our no-pooling estimate
of theta_i is the ratio of the observed count values of interest y_i to
the reference count values n_i.
}
