% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{select_for_report}
\alias{select_for_report}
\title{Select data for report}
\usage{
select_for_report(
  d,
  unit_df = NULL,
  unit_group_name = NULL,
  min_y_new = 0,
  min_diff = 0
)
}
\arguments{
\item{d}{Data frame as returned from \emph{countcheck()}}

\item{unit_df}{Data frame with columns
\emph{unit}, \emph{unit_group_name}, \emph{unit_name},
and \emph{unit_url} (ignored if NA).
Must contain exactly one record for each unit in
\emph{d}.
(default: NULL)}

\item{unit_group_name}{Name of unit group for which data
should be returned. Can be part of name (substring).
(default: NULL)}

\item{min_y_new}{Minimum value for y_new (default: 0)}

\item{min_diff}{Minimum difference between y_new and ucl_partpool
(default: 0)}
}
\value{
Data frame to be passed in \emph{countcheck_list}
  to \emph{html_report()}
}
\description{
Selects data for report.
}
\details{
Selected columns:
\itemize{
\item \emph{unit}:            ID for unit
\item \emph{y_new}:           new count values of interest
\item \emph{ucl_partpool}:    UCL based on n_new and partial-pooling
                                estimate of theta
\item \emph{fe_partpool}:     factor \emph{f}, with observed y_new
                                exceeding partial-pooling UCL by
                                \emph{f * sd(y_new)},
                                given n_new and partial-pooling estimate
                                of theta
}

Condition that always applies: \emph{y_new} must exceed \emph{ucl_partpool}.

Additional optional conditions:
\itemize{
\item \emph{y_new} exceeds \emph{min_y_new} (default: 0)
\item difference between \emph{y_new} and \emph{ucl_partpool}
  exceeds \emph{min_diff} (default: 0)
\item \emph{unit_group_name} (if present) matches part of group name
  in \emph{unit_df} (if present) for units in \emph{d}
}

Returned data is sorted by \emph{fe_partpool} in descending order.
Therefore, items towards the top are less likely to be the result of
random fluctuations.
}
