% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_data}
\alias{simulate_data}
\title{Simulate data for use in examples}
\usage{
simulate_data(units = 1000, random_seed = NULL)
}
\arguments{
\item{units}{Number of observational units (default: 1000)}

\item{random_seed}{Seed value (optional)}
}
\value{
data frame initialized with simulated data
}
\description{
The distributions and parameters in this simulation have been chosen
so that the generated data is roughly comparable to certain data of
interest.
}
