% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{theta_nopool}
\alias{theta_nopool}
\title{No-pooling model}
\usage{
theta_nopool(n, y)
}
\arguments{
\item{n}{Previous reference count values (measure of exposure),
must at least be 1}

\item{y}{Previous count values of interest}
}
\value{
No-pooling estimates of \emph{theta}
}
\description{
For the no-pooling model we estimate \emph{theta_i} using only the observed
counts of observational unit \emph{i}. For each unit \emph{i} our no-pooling
estimate of \emph{theta_i} is the ratio of the observed count values of
interest \emph{y_i} to the reference count values \emph{n_i}.
}
