% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucls.R
\name{ucl}
\alias{ucl}
\title{Calculate upper control limit (UCL)}
\usage{
ucl(theta_hat, n_new, factor_sd = 3)
}
\arguments{
\item{theta_hat}{Estimate (unless known) of rate parameter \emph{theta}}

\item{n_new}{New value for exposure (reference counts)}

\item{factor_sd}{Factor multiplying standard deviation (default: 3)}
}
\value{
UCLs
}
\description{
Basic formula for UCL:
\eqn{estimated mean + 3 * estimated standard deviation}
}
\details{
In this case, assuming a Poisson distribution, both the estimated mean
and the estimated variance are
\deqn{lambda_hat = theta_hat * n_new}

Rounds the result to the nearest integer.

Adds 0.5, so that count values are either above or below the UCL
(as in Wheeler & Chambers [1992]).
}
