% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{add_theta_complpool}
\alias{add_theta_complpool}
\title{Add complete-pooling estimate of \emph{theta} to data frame}
\usage{
add_theta_complpool(d)
}
\arguments{
\item{d}{Initialized data frame}
}
\value{
Data frame with values for \emph{theta_complpool}
}
\description{
Adds complete-pooling estimate of \emph{theta} to data frame.
}
\keyword{internal}
