% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{add_theta_nopool}
\alias{add_theta_nopool}
\title{Add no-pooling estimate of theta to data frame}
\usage{
add_theta_nopool(d)
}
\arguments{
\item{d}{Initialized data frame}
}
\value{
data frame with values for theta_nopool
}
\description{
Add no-pooling estimate of theta to data frame
}
