% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{add_theta_partpool}
\alias{add_theta_partpool}
\title{Add partial-pooling estimate of \emph{theta} to data frame}
\usage{
add_theta_partpool(d, random_seed = 200731)
}
\arguments{
\item{d}{Initialized data frame}

\item{random_seed}{Seed value for Stan (default: 200731)}
}
\value{
Data frame with values for \emph{theta_partpool}
}
\description{
Add partial-pooling estimate of \emph{theta} (based on Bayesian hierarchical
model) to data frame.
}
