% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucls.R
\name{add_ucl_complpool}
\alias{add_ucl_complpool}
\title{Add complete-pooling UCL to data frame}
\usage{
add_ucl_complpool(d)
}
\arguments{
\item{d}{Initialized data frame}
}
\value{
data frame with values for ucl_complpool
}
\description{
UCL based on n_new and complete-pooling estimate of theta
}
