% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucls.R
\name{add_ucl_nopool}
\alias{add_ucl_nopool}
\title{Add no-pooling UCL to data frame}
\usage{
add_ucl_nopool(d)
}
\arguments{
\item{d}{Initialized data frame}
}
\value{
data frame with values for ucl_nopool
}
\description{
UCL based on n_new and no-pooling estimate of theta
}
