% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucls.R
\name{add_ucl_partpool}
\alias{add_ucl_partpool}
\title{Add partial-pooling UCL to data frame}
\usage{
add_ucl_partpool(d)
}
\arguments{
\item{d}{Initialized data frame}
}
\value{
data frame with values for \emph{ucl_partpool}
}
\description{
Add UCL based on \emph{n_new} and partial-pooling estimate of \emph{theta}
to data frame.
}
