% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucls.R
\name{add_ucl_partpool}
\alias{add_ucl_partpool}
\title{Add partial-pooling UCL to data frame}
\usage{
add_ucl_partpool(d, factor_sd = 3)
}
\arguments{
\item{d}{Initialized data frame}

\item{factor_sd}{Factor multiplying standard deviation (default: 3)}
}
\value{
Data frame with values for \emph{ucl_partpool}
}
\description{
Adds UCL based on \emph{n_new} and partial-pooling estimate of \emph{theta}
to data frame.
}
\keyword{internal}
