% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucls.R
\name{add_ucl_true_theta}
\alias{add_ucl_true_theta}
\title{Add "true" UCL to data frame}
\usage{
add_ucl_true_theta(d, factor_sd = 3)
}
\arguments{
\item{d}{Initialized data frame}

\item{factor_sd}{Factor multiplying standard deviation (default: 3)}
}
\value{
Data frame with values for \emph{ucl_true_theta}
}
\description{
Adds UCL based on \emph{n_new} and true value of \emph{theta} to
data frame.
}
\keyword{internal}
