% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucls.R
\name{add_ucl_true_theta}
\alias{add_ucl_true_theta}
\title{Add "true" UCL to data frame}
\usage{
add_ucl_true_theta(d)
}
\arguments{
\item{d}{Initialized data frame}
}
\value{
data frame with values for \emph{ucl_true_theta}
}
\description{
Add UCL based on \emph{n_new} and true value of \emph{theta} to
data frame.
}
