% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{html_report}
\alias{html_report}
\title{Generate HTML report}
\usage{
html_report(
  countcheck_list,
  unit_df,
  title = "Report",
  table_width_px = 1000,
  column_headers = c(group = "Group", count = "Count", ucl = "UCL", unit = "Unit", name =
    "Name"),
  charset = "utf-8",
  lang = "en",
  home_url = NULL,
  style = NULL
)
}
\arguments{
\item{countcheck_list}{Data to be shown in report. List of lists with
two items named "caption" and "countcheck_df".
The item \emph{countcheck_df} in each of the nested lists is a data frame
with columns \emph{y_new}, \emph{ucl_partpool}, and \emph{unit} -
as returned by \emph{select_for_report()}.
Each \emph{countcheck_df} must contain at least one row.
The item \emph{caption} in each of the nested lists is used for
the table caption.}

\item{unit_df}{Data frame with columns
\emph{unit}, \emph{unit_group_name}, \emph{unit_name},
and \emph{unit_url} (ignored if NA).
Must contain exactly one record for each distinct unit in
\emph{countcheck_df} dataframes in \emph{countcheck_list}.}

\item{title}{Title for HTML document}

\item{table_width_px}{Table width in pixels (for all tables)}

\item{column_headers}{Column headers for tables; vector with five headers -
the elements of the vector must be named \emph{group}, \emph{count},
\emph{ucl}, \emph{unit}, and \emph{name}}

\item{charset}{Character set of data and HTML document
(must contain ASCII as subset)}

\item{lang}{Language of HTML document}

\item{home_url}{URL for link to "Home" included at top of HTML document
(optional)}

\item{style}{Additional CSS rules (optional)}
}
\value{
String with HTML code for report
}
\description{
Generates HTML report for multiple tables as returned by
\emph{select_for_report()}.
}
