% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{html_table}
\alias{html_table}
\title{Generate HTML table}
\usage{
html_table(
  countcheck_df,
  unit_df,
  caption,
  headers = c(group = "Group", count = "Count", ucl = "UCL", unit = "Unit", name =
    "Name")
)
}
\arguments{
\item{countcheck_df}{Data frame with columns
\emph{y_new}, \emph{ucl_partpool}, \emph{unit}}

\item{unit_df}{Data frame with columns
\emph{unit}, \emph{unit_group_name}, \emph{unit_name}, \emph{unit_url}}
}
\value{
String with HTML code for table
}
\description{
Generates HTML table.
}
\keyword{internal}
