% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{initialize}
\alias{initialize}
\title{Initialize data frame}
\usage{
initialize(unit, n, y, n_new, y_new, true_theta = NULL)
}
\arguments{
\item{unit}{ID for unit}

\item{n}{Previous reference count values (measure of exposure)}

\item{y}{Previous count values of interest}

\item{n_new}{New reference count values}

\item{y_new}{New count values of interest}

\item{true_theta}{True value of theta (if known; optional)}
}
\value{
Initialized data frame
}
\description{
Construct data frame with these fields:
\itemize{
\item \emph{unit}:            ID for unit
\item \emph{n}:               previous reference count values
                              (measure of exposure)
\item \emph{y}:               previous count values of interest
\item \emph{n_new}:           new reference count values
\item \emph{y_new}:           new count values of interest
\item \emph{true_theta}:      true value of theta (if known)
\item \emph{theta_nopool}:    no-pooling estimate of theta
\item \emph{theta_complpool}: complete-pooling estimate of theta
\item \emph{theta_partpool}:  partial-pooling estimate of theta
\item \emph{ucl_true_theta}:  UCL based on n_new and true value of theta
                              (if known)
\item \emph{ucl_nopool}:      UCL based on n_new and no-pooling estimate
                              of theta
\item \emph{ucl_complpool}:   UCL based on n_new and complete-pooling
                              estimate of theta
\item \emph{ucl_partpool}:    UCL based on n_new and partial-pooling
                              estimate of theta
}
Input vectors must be of equal length, with length >= 3.
\emph{unit}, \emph{n}, \emph{y}, \emph{n_new}, \emph{y_new},
and \emph{true_theta} (if known) are initialized using the
corresponding parameters. The other fields (and \emph{true_theta}
(if not known)) are initialized with NA.
}
