% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{select_for_report}
\alias{select_for_report}
\title{Select data for report}
\usage{
select_for_report(d)
}
\arguments{
\item{d}{Data frame as returned from \emph{countcheck()}}
}
\value{
Data frame to be passed in \emph{countcheck_list}
  to \emph{html_report()}
}
\description{
Selects data for report.
}
