% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{theta_complpool}
\alias{theta_complpool}
\title{Complete-pooling model}
\usage{
theta_complpool(n, y)
}
\arguments{
\item{n}{Previous reference count values (measure of exposure)}

\item{y}{Previous count values of interest}
}
\value{
Complete-pooling estimates of theta
}
\description{
For the complete-pooling model we consider theta_i to be the same
for all observational units. We assume that observed data can be
treated as coming from a homogeneous source. For all units i our
complete-pooling estimate of theta_i is the ratio of the sum of
the observed count values of interest y_i to the sum of the
reference count values n_i.
}
