% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucls.R
\name{ucl}
\alias{ucl}
\title{Calculate upper control limit (UCL)}
\usage{
ucl(theta_hat, n_new)
}
\arguments{
\item{theta_hat}{Estimate (unless known) of rate parameter theta}

\item{n_new}{New value for exposure (reference counts)}
}
\value{
UCL
}
\description{
Basic formula for UCL: estimated mean + 3 * estimated standard deviation
}
\details{
In this case, assuming a Poisson distribution, both the estimated mean
and the estimated variance are lambda_hat = theta_hat * n_new.

Round the result to the nearest integer.

Add 0.5, so that count values are either above or below the UCL
(as in Wheeler & Chambers [1992]).
}
